+++ src/transaction_builder.js
@@ -6,6 +6,7 @@ import xdr from './generated/stellar-xdr_generated';
 import { Keypair } from './keypair';
 import { Transaction } from './transaction';
 import { Memo } from './memo';
+import { Network } from './network';
 
 /**
  * Minimum base fee for transactions. If this fee is below the network
@@ -52,18 +53,18 @@ export const TimeoutInfinite = 0;
  * a payment to `destinationB`. The built transaction is then signed by `sourceKeypair`.</p>
  *
  * ```
- * var transaction = new TransactionBuilder(source, { fee })
- *  .addOperation(Operation.createAccount({
-        destination: destinationA,
-        startingBalance: "20"
-    })) // <- funds and creates destinationA
-    .addOperation(Operation.payment({
-        destination: destinationB,
-        amount: "100",
-        asset: Asset.native()
-    })) // <- sends 100 XLM to destinationB
- *   .setTimeout(30)
- *   .build();
+ * var transaction = new TransactionBuilder(source, { fee, networkPassphrase: Networks.TESTNET })
+ * .addOperation(Operation.createAccount({
+ *     destination: destinationA,
+ *     startingBalance: "20"
+ * })) // <- funds and creates destinationA
+ * .addOperation(Operation.payment({
+ *     destination: destinationB,
+ *     amount: "100",
+ *     asset: Asset.native()
+ * })) // <- sends 100 XLM to destinationB
+ * .setTimeout(30)
+ * .build();
  *
  * transaction.sign(sourceKeypair);
  * ```
@@ -75,6 +76,7 @@ export const TimeoutInfinite = 0;
  * @param {number|string|Date} [opts.timebounds.minTime] - 64 bit unix timestamp or Date object
  * @param {number|string|Date} [opts.timebounds.maxTime] - 64 bit unix timestamp or Date object
  * @param {Memo} [opts.memo] - The memo for the transaction
+ * @param {string} [opts.networkPassphrase] passphrase of the target stellar network (e.g. "Public Global Stellar Network ; September 2015").
  */
 export class TransactionBuilder {
   constructor(sourceAccount, opts = {}) {
@@ -93,6 +95,7 @@ export class TransactionBuilder {
     this.timebounds = clone(opts.timebounds) || null;
     this.memo = opts.memo || Memo.none();
     this.timeoutSet = false;
+    this.networkPassphrase = opts.networkPassphrase || null;
   }
 
   /**
@@ -161,6 +164,17 @@ export class TransactionBuilder {
   }
 
   /**
+   * Set network nassphrase for the Transaction that will be built.
+   *
+   * @param {string} [networkPassphrase] passphrase of the target stellar network (e.g. "Public Global Stellar Network ; September 2015").
+   * @returns {TransactionBuilder}
+   */
+  setNetworkPassphrase(networkPassphrase) {
+    this.networkPassphrase = networkPassphrase;
+    return this;
+  }
+
+  /**
    * This will build the transaction.
    * It will also increment the source account's sequence number by 1.
    * @returns {Transaction} This method will return the built {@link Transaction}.
@@ -211,7 +225,7 @@ export class TransactionBuilder {
     xtx.operations(this.operations);
 
     const xenv = new xdr.TransactionEnvelope({ tx: xtx });
-    const tx = new Transaction(xenv);
+    const tx = new Transaction(xenv, this.networkPassphrase);
 
     this.source.incrementSequenceNumber();